Added yes-no option to uninstall

+++ install/uninst.sh
@@ -109,23 +109,26 @@ function yesno() {
 
 # Start the uninstall
 echo; sleep 1
-if [[ -f /usr/local/bin/stir ]]; then
-  echo "Removing stir"
-  sudo rm /usr/local/bin/stir; error_check
 
-  if [[ -d /etc/stir/lib ]]; then
-    sudo rm -rf /etc/stir/lib; error_check
-  fi
+if yesno --default yes "Uninstall stir? [Y/n] "; then
+  if [[ -f /usr/local/bin/stir ]]; then
+    echo "Removing stir"
+    sudo rm /usr/local/bin/stir; error_check
 
-  if [[ -d /etc/stir/html ]]; then
-    sudo rm -rf /etc/stir/html; error_check
-  fi
-
-  if yesno --default yes "Remove configuration files? [Y/n] "; then
-    if [[ -d /etc/stir ]]; then 
-      sudo rm -rf /etc/stir; error_check
+    if [[ -d /etc/stir/lib ]]; then
+      sudo rm -rf /etc/stir/lib; error_check
     fi
+
+    if [[ -d /etc/stir/html ]]; then
+      sudo rm -rf /etc/stir/html; error_check
+    fi
+
+    if yesno --default yes "Remove configuration files? [Y/n] "; then
+      if [[ -d /etc/stir ]]; then 
+        sudo rm -rf /etc/stir; error_check
+      fi
+    fi
+    echo "Successfully uninstalled."
+    exit 0
   fi
-  echo "Successfully uninstalled."
-  exit 0
 fi
